rename test utility
